feat(api): add FluidStack utility
This object is capable of tracking the internal contents of a pipette,
differentiating between liquid and air.

This is a key capability because we need to be able to say how much
_liquid_ we just put into a well when we dispensed, independently of how
much _air_ (from previous air gaps) we may have also ejected. Since
people may add air to their pipettes either above (because they want
some elbow room to overdispense) or below (because they want to avoid
drips) or even conceptually between (because it's not impossible to do)
liquid, we also need to know the relative location of the air and the
liquid.

To do this we introduce a new concept of "fluid" as "air or liquid";
really this is "anything that flows", but whatever. Pipettes contain
fluid, and state stores will update how much fluid they contain. We
store a stack (formally - LIFO) of fluids by type and volume; if you
add fluid of the same kind that's at the bottom, we'll add to the
tracked volume and if you add fluid of a different kind, we'll add a
stacked element. It's the same thing in reverse for removing fluid.

The ultimate key, however, is the function
liquid_part_of_dispense_volume(), which answers the question posed
above: if I dispense X ul of whatever's currently in my pipette, how
much liquid will I deposit in the well?