refactor(api): track liquid in pipettes
Today, we store the contents of a pipette as an optional float. If the
value is None, then the pipette has an unknown amount of fluid in it and
can't add or remove any; if the value is not None, then the pipette has
that many ul of fluid in it. If we dispense, we dispense that much liquid
into a well.

This isn't good enough to accurately track liquid, because air gaps
exist. We need to separately track how much liquid and how much air is
in a pipette, and figure out how much of each we're dispensing when we
dispense somewhere.

To do that,
- Swap out the aspirated_volume_by_id: Dict[str, Optional[float]] with
pipette_contents_by_id: Dict[str, Optional[FluidStack]]
- Swap out volume tracking from success/fail action based tracking in
the stores to state update based setting from the commands
- Swap out dispense(inPlace)? blindly adding the volume it dispensed to a target
well with dispense(inPlace)? asking the store how much liquid it
dispensed and adding that to the well

This ends up being a tremendous amount of code because basically all of
the tests for everything have to change, which is annoying. There's
another step that should happen here where the state tests get
simplified a lot but this was already huge and it's already very
important that the underlying logic didn't change, so let's keep the
diffs (relatively) small.