refactor(api): control nozzlemap type better
There's this hardware controller NozzleMap type that is mostly internal
but actually exposed upstream in a couple weird uncontrolled ways.
Refactor this so that
- There's a controlled interface that is in opentrons.types and that is
the only thing that is exposed above the engine
- The engine and lower are allowed to see the actual type

This had some knock-on consequences because some functionality had to
move to lower layers. Specifically, "can this layout support LLD" is a
question that now only the engine can answer because the physical
configuration type is not in the interface, so push it down (which feels
better anyway because now you get this checked if you use the commands api).