feat(api): add labware-scope liquid loading api
We had the ability to load liquids on individual Wells, but this creates
an engine command per well - which can be very onerous for large labware
- and requires loops and such to load liquids. The common case for
loading liquids is that everything is the same, and the second common
case is that you care about a labware as a whole at a time (e.g. because
you're reading from a platemap).

We can support these uses better by having Labware.load_liquid (for many
wells with the same amount of liquid), Labware.load_liquid_by_well (more
complex, for more complex use cases) and Labware.load_empty, all taking
arguments of the type that you can derive from other PAPI methods or
domain-specific reasoning.

Also, remove well.load_empty because it wasn't shipped yet and is now
duplicative.

Closes EXEC-825