feat(api): implement PE error monad
Let's make an intermediate layer of command execution, inside the
command domain so we don't have to move a bunch of types, that calls
execution layer functions and returns a new Maybe class that has nice
monadic chaining calls like a JS promise or some other burrito-style
implementations.

As an example, let's use it for overpressure errors in PrepareToAspirate.

Closes EXEC-830