fix(components, protocol-designer): fix labware thermocycler render issue
Because protocol designer is unaware of actual thermocycler lid state until a thermocycler step
explicitly sets its state, we default to a gray box. This PR resets that logic to show the lid as
open if the lid state is indeterminate to avoid confusion; however, error messages prompting the
user to open the lid before moving labware/liquid to or from the thermocycler will persist, as the
thermocycler lid is neither open nor closed. Also, this fixes a rendering issue where labware was
shown on top of the closed thermocycler lid.

Closes RQA-3556