refactor(api): handle failures straightforwardly
Unfortunately, the Maybe experiment fails when you want to do things
like chain async methods. However, the helpers added to StateUpdate
still make this a bit less of a painful process, and even after
refactoring the microoperations back to resulting in unions I think it's
a net improvement.

There's still value in the micro-operation path, so extend that across
the rest of the pipetting verbs and add a new one for move_to_well.