fix(app, robot-server): support `retryLocation` when retrying `dropTipInPlace` during Error Recovery (#16839)
Closes RQA-3591

When dropping a tip in a trash bin or waste chute, the pipette moves downwards before doing the drop tip. During Error Recovery, when we retry the failed command with a fixit intent, we need to perform a moveToCoordinates first before dispatching the dropTipInPlace or the robot drops the tip(s) from too high a location.

To know how far to move, we need the failedCommand to contain retryLocation metadata. We have a pattern for this already with overpressure in place commands, so we extend the same functionality here.

Co-authored-by: Max Marrone <max@opentrons.com>