fix(protocol-designer): fix layout for ProtocolSteps
This PR fixes the layout for protocol steps, acocunting for TimelineAlerts. To maintain position
without flickering when hovering timeline steps, we align the center content to the top of the
container, and provide top padding dependent on the presence of timeline alerts. We also fix logic
to not show timeline alerts if a stepform toolbox is open. Lastly, we make the container scrollable
to handle timeline alerts and step summaries rendering at the same time, and set the timeline,
stepform, and substeps toolbox heights to fill the view height with padding.