fix(components, protocol-designer): fix logic for disabling advanced transfer settings (#16996)
Update logic for disabling mix/blowout settings depending on transfer
path.
- If path is consolidate, aspirate mix field should be disabled.
- If path is distribute, dispense mix field should be disabled, and
blowout field should be disabled if disposal volume field is enabled.

Side effect of enabling disposal volume field is to disable and collapse
blowout field. These changes implicate that Check and ListItem
components accommodate disabled state.

Note that I don't love this fix— it breaks pattern with how we handle
tooltips, but it looks like we haven't had to deal with multiple
disabled tooltips on the same form field in the past. I am open to ideas
on how to migrate this to a more sustainable pattern moving forward.