fix(app): fix inescapable pipette wizard (#17008)
Closes RABR-675

It's possible to sometimes get stuck in the pipette wizard without an exit button if the error occurs while a command is executing or the maintenance run is trying to be deleted. Additionally (and I think this is how the bug ABR reported happens to begin with) - if an error does occur once during pipette flows, we were just closing the wizard without deleting the maintenance run. handleCleanUpAndClose conditionally closes the run if it does exist, which we definitely want to do if we error.

Lastly, it's probably better to show an exit button while certain requests are pending and just disable it rather than avoid rendering an exit button altogether.

This control flow more closely follows what happens in GripperWizardFlows, which is probably why ABR & QA haven't seen this bug crop up in that flow.