refactor(robot-server): use different pipette IDs for OT2 vs Flex dev servers (#17031)
# Overview

The OT2 and Flex pipette calibration data format is slightly different.
In the simulator setup file, we have the OT2 & Flex pipette IDs the same
so the calibration files get unintentionally shared between OT2 & Flex.
When switching between OT2 and Flex dev servers, the difference in data
format causes a validation error and we have to re-calibrate the
pipettes (or update the files) each time. It's quite annoying to have to
do this so I am changing the Flex's pipette ID so that we'll be able to
use separate calibration files for flex & OT2 pipettes.

## Risk assessment

None. Dev server change only