fix(protocol-designer): fix logic for showing AutoAddPauseUntilTempStepModal
The logic for `isPristineSetTempForm` passed to `StepForm` was performing a brittle check on the
state of the temperature module form. Here, I change the logic to check for the presence of a
non-null targetTemperature for safety, more closely mirroring the functioning logic of
`isPristineSetHeaterShakerTempForm`. This fix correctly presents the modal to add a pause step after
setting temperature for temperature modules.

Closes RQA-3763