fix(hardware,api): remove hardware import from simulate
opentrons_simulate was importing from opentrons_hardware via the
hardware controller protocol interface, which was doing it to support
sensor data queue spying.

Fixing this required a pretty big refactor to make the link between the
controller and opentrons_hardware be a callback that could reformat
data, so we can make versions of the relevant firmware binding types
that are safe for reexport.

This work should be extended to the capacitive sensor queue at some point.

Closes RQA-3766