chore(monorepo): Update Electron to v33.2.1 (#17010)
Closes EXEC-1032

This PR continues the work started in #15894, cherry-picking several commits and completes updating Electron to 33.2.1. There are a couple dependency updates as well: node gets a big bump to 22.11.0,electron-rebuild bumps to 3.7.1, and electron-builder bumps to 25.1.8. This PR does not bump USB-related dependencies, since USB works as is (and if we're going to bump those dependencies, it would be nice to do them in an isolated PR).