Remove query parameters from DELETE endpoint.
I don't think there was anything fundamentally wrong with this, and in fact it seemed kind of nice that you could just change a GET to a DELETE just by flipping the method. But we would want to fully mirror GET and DELETE, and putting the cursor/pageLength parameters in the DELETE endpoint to do that felt...weird.