Shift datetime workaround from production code to tests.
Pydantic serializes datetimes with a "Z", which datetime.datetime cannot parse in Python 3.10 (but can in Python 3.11). Instead of working around this by changing how Pydantic serializes datetimes, this works around it by fixing up the parsing in the tests.

This will be easier to stay on top of--we won't have to adjust every datetime field in every one of our Pydantic models. It will also be easier to remove in Python 3.11.