Fix robot-server docs build script.
The actual problem was passing an int where a str was expected.

This was failing in a way that was more annoying than necessary to debug because we were only showing the exception message, not the stack trace. Since this is an internal tool for dev and CI consumption, we have no reason to hide stack traces, so this lets the exception propagate and display in the default Python way.