feat(protocol-designer, app, api): step grouping foundation (#15737)
addresses AUTH-573

# Overview

co-authored by @jbleon95 

This is the foundation for Step grouping in the system. Step grouping is
part of the PD redesign follow-up (originally it was part of the PD
redesign but got descoped a few days ago. it IS still part of the
redesign plans but will be in a follow up)

The way step grouping works is the user in PD adds the step to a group
and that information gets parsed into `commandAnnotations` in the schema
as part of a `secondOrderCommand`.

When uploading the protocol to the app, the `commandAnnotations` is part
of the `CompletedProtocolAnalysis` which is displayed in the Timeline
tab in `ProtocolDetails` - only on the desktop app. Note that the
Timeline tab is only visible if you have the `Protocol timeline` feature
flag turned on.

---------

Co-authored-by: jbleon95 <jeremybenleon@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: jbleon95 <12025228+jbleon95@users.noreply.github.com>