fix(protocol-designer): handle duplicate labware with a full deck (#17083)
If a slot overflow is opened and duplicate labware is clicked, the
action will fail silently if no room is left on the robot deck. Here, we
use the `getNextAvailableDeckSlot` util to determine whether or not we
can dispatch a labware duplication on the deck. If not, we keep the menu
open, but render a snackbar letting the user know the reason for the
failure. Note that according to new design paradigms, the button remains
enabled in every case.