fix(protocol-designer): blowout field checkbox properly populating (#17093)
closes RQA-3792

This was a weird bug only affecting the UI and not any commands being
generated. Basically, the blowout checkbox was incorrectly not being
selected because there was logic in place to hardcode it
to false due to the disposal volume checkbox being true. IDK why the
disposal volume checkbox is true -- seems like something we allowed for
all move liquid steps back in the day. So for a UI fix, I added the
field to be dependent on the disposal volume fields and removed the
useEffect in place.