Revert use of pydantic_core.from_json() in labware def validation.
from_json() appears not to be *exactly* a drop-in replacement for json.loads(): it raises different exceptions for invalid input. We could certainly account for that, but for better testing and reviewing, it seems like it should be in a different PR at this point.

I'm not reverting the multi-schema additions to this function because there doesn't seem to be a problem with those.