refactor(app): remove unneeded source maps dependency (#17100)
After playing with source map dependencies, it turns out that we don't need the extra library, and the tsconfig source maps are irrelevant due to the way vite bundles source maps. All we need is to keep the existing sourceMaps: true property in our vite.config build options! This cuts the bundle size dramatically: it's now only ~3MB increase.

This also fixes an interaction with the source maps library and the network settings tab, causing the app to crash (on the desktop app).