fix(app): fix dev ODD system image build causing infinite spinner (#17104)
Dev ODD system image builds have been failing. When digging into the Electron renderer process logs, this appears to be because of the way we inject the mixpanel config in our dev builds: it's not what mixpanel expects. During mixpanel initialization, if an unexpected configuration/id is supplied, certain methods on mixpanel are undefined. We've had similar issues like this in the past, and a more general-case solution to all of these issues is just to wrap the whole mixpanel initialization process in a try-catch block.