Fix text decoding when parsing JSON from files.
Formerly, Pydantic had `Model.parse_file()`, where the encoding is configurable but defaults to UTF-8. Earlier commits in this PR replaced that with `pathlib.Path.read_text()`, where the encoding is configurable but defaults to something platform-dependent, which is UTF-16 on Windows. So on Windows, we were accidentally trying to read files as UTF-16 where in fact they were, apparently, UTF-8 or ASCII, and this caused various parse errors.

The fix here is to read the file as binary, instead of text, so we can avoid saying anything about the encoding. Figuring out the encoding then becomes the problem of the underlying JSON library, which can use JSON-specific heuristics to autodetect it. Python's default `json.load()`, for example, tolerates UTF-8, UTF-16, and UTF-32.