fix(protocol-designer): make 8_2_2 migration to migrate HS set timer … (#17124)
…field

The `heaterShakerSetTimer` field's type is `heaterShakerSetTimer: 'true'
| 'false' | null` - this was written 3 years ago. However, despite this,
the field when checking the checkbox in the form in PD would populate it
as `true` or `false` booleans rather than the true/false strings.

2 weeks ago, right before 8.2.0 was released, there was a bug checking
the timer checkbox for the temperature module and it was because of this
boolean vs boolean strings issue. So I extended it onto the
heater-shaker set timer field as well because of its type. Unfortunately, that resulted in this
current error:

- protocols created in 8.2.0 and 8.2.1 have the `heaterShakerSetTimer`
field set to either `'true'` or `'false'` strings
- protocols created prior to 8.2.0 have the `heaterShakerSetTimer` field
set to `true` or `false` boolean
- my initial fix was to update the form field to check for `true` or `false` boolean but
that didn't fix any protocols created in 8.2.0 and 8.2.1.
- the extended fix unfortunately required a migration so now if `'true'`
or `'false'` strings are used in the field, they get migrated to `true`
or `false` boolean -- this i believe is the only way to fix all protocol
versions