feat(shared-data,-protocol-designer): add foundation for plate reader support in PD
This PR introduces plate reader to PD, still hidden behind a feature flag. Namely, it adds the ability to add plate reader to the initial deck state and fixes logic for ModuleLabel rendering for all modules. Of note, in this dev work, I discovered a bug in the plate reader's shared-data definition, in which the top-level x and y dimensions were flipped, so those values are swapped here as well.