fix(hardware): do not busyloop if driver isnt real
If a can messenger was built wrapping a can driver that immediately
completed (because, say, it was a mock) the read task would busyloop
without yielding. Add a scheduler yield to prevent this and log when it
happens (we are also making the mock more realistic to avoid triggering
the log).