chore(api): fix can messenger warnings
The can messenger used in the controller task was emitting warnings for
every test about not being cleaned up properly. This is because it was
not being cleaned up properly.

However, cleaning it up properly requires also making it work properly.
We need to have a real(ish) implementation for the async-iterator
interface of the mock driver (aka no message received, so never yield)
or we busy loop.